+++ game-avowed/CHANGELOG.md
@@ -6,5 +6,7 @@
 - Added: Notification indicating deployment is required after changing the load order.
 - Fixed: Missing FOMOD installer check for pak mods.
 - Fixed: path strings
-- Added: Buttons to open PCGamingWiki page, view changelog, and submit bug reports
 - Fixed: Xbox game version detection
+- Added: UE4SS DLL Mods installer
+- Added: Button to download UE4SS and open PCGamingWiki page, view changelog, and submit bug reports
+- Added: Automatic "Win64" to "WinGDK" folder rename for Xbox version UE4SS combo mod installer